+++ src/system/helpers.h
@@ -47,7 +47,7 @@ std::string trim(std::string &str, const std::string &trimChars = " \n\r\t");
 class CFileHelpers
 {
 	private:
-		int FileBufSize;
+		unsigned long FileBufSize;
 		char *FileBuf;
 		int fd1, fd2;
